reorganize attacks
